Database now has the data. The Get() method receives the data and the programs still work with it. Get(int id) is half changed, unsure if still needed (never has been used, asked Kenji for clarification). Search(string query) hasn't been changed at all.

+++ WebstorePhones.Domain/Objects/Phone.cs
@@ -4,13 +4,13 @@ namespace WebstorePhones.Domain.Objects
 {
     public class Phone
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Brand { get; set; }
         public string Type { get; set; }
         public string FullName { get { return $"{Brand} - {Type}"; } }
         public string Description { get; set; }
-        public double PriceWithTax { get; set; }
-        public double PriceWithoutTax { get { return Math.Round(PriceWithTax / (1 + 0.21), 2); } }
+        public decimal PriceWithTax { get; set; }
+        public decimal PriceWithoutTax { get { return Math.Round(PriceWithTax / (1 + 0.21m), 2); } }
         public int Stock { get; set; }
     }
 }
